+++ examples/react/e-commerce-with-chat/App.mobile.css
