+++ examples/react/e-commerce-with-chat/App.tsx
@@ -41,12 +41,130 @@ import './App.css';
 import './components/Pagination.css';
 import './App.mobile.css';
 
+import type { ChatMessageBase } from 'instantsearch-ui-components';
 import type { Hit as AlgoliaHit } from 'instantsearch.js';
 
 const searchClient = algoliasearch(
   'latency',
   '6be0576ff61c053d5f9a3225e2a90f76'
 );
+
+const AssistantLeadingComponent = () => (
+  <div
+    style={{
+      width: '24px',
+      height: '24px',
+      borderRadius: '9999px',
+      backgroundColor:
+        'rgba(var(--ais-primary-color-rgb), var(--ais-primary-color-alpha))',
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+      color:
+        'rgba(var(--ais-button-text-color-rgb), var(--ais-button-text-color-alpha))',
+      fontSize: '12px',
+      fontWeight: 'bold',
+    }}
+  >
+    A
+  </div>
+);
+
+const UserLeadingComponent = () => (
+  <div
+    style={{
+      width: '24px',
+      height: '24px',
+      borderRadius: '9999px',
+      backgroundColor: 'rgba(var(--ais-muted-color-rgb), 0.5)',
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+      color:
+        'rgba(var(--ais-button-text-color-rgb), var(--ais-button-text-color-alpha))',
+      fontSize: '12px',
+      fontWeight: 'bold',
+    }}
+  >
+    U
+  </div>
+);
+
+const copyIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="16"
+    height="16"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <rect width="14" height="14" x="8" y="8" rx="2" ry="2" />
+    <path d="M4 16c-1.1 0-2-.9-2-2V4c0-1.1.9-2 2-2h10c1.1 0 2 .9 2 2" />
+  </svg>
+);
+
+const regenerateIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="16"
+    height="16"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <path d="M3 12a9 9 0 0 1 9-9 9.75 9.75 0 0 1 6.74 2.74L21 8" />
+    <path d="M21 3v5h-5" />
+    <path d="M21 12a9 9 0 0 1-9 9 9.75 9.75 0 0 1-6.74-2.74L3 16" />
+    <path d="M8 16H3v5" />
+  </svg>
+);
+
+const copyToClipboard = (message: ChatMessageBase) => {
+  navigator.clipboard.writeText(
+    message.parts
+      .map((part) => {
+        if ('text' in part) {
+          return part.text;
+        }
+        return '';
+      })
+      .join('')
+  );
+};
+
+const assistantActions = [
+  {
+    title: 'Copy to clipboard',
+    icon: copyIcon,
+    onClick: copyToClipboard,
+  },
+  {
+    title: 'Regenerate',
+    icon: regenerateIcon,
+    onClick: () => {
+      // TODO: regenerate functionality would need to be passed down from Chat component
+    },
+  },
+  {
+    title: 'More actions',
+    disabled: true,
+  },
+];
+
+const userActions = [
+  {
+    title: 'Copy to clipboard',
+    icon: copyIcon,
+    onClick: copyToClipboard,
+  },
+];
 
 const indexName = 'instant_search';
 const routing = getRouting(indexName);
@@ -111,7 +229,17 @@ export function App() {
       <Chat
         agentId="5bc03a4f-4e25-400c-9f52-a7b969d7f3da"
         itemComponent={ItemComponent}
-        messagesProps={messagesProps}
+        messagesProps={{
+          ...messagesProps,
+          assistantMessageProps: {
+            actions: assistantActions,
+            leadingComponent: AssistantLeadingComponent,
+          },
+          userMessageProps: {
+            actions: userActions,
+            leadingComponent: UserLeadingComponent,
+          },
+        }}
       />
       <header className="header" ref={headerRef}>
         <p className="header-logo">
