+++ examples/react/e-commerce-with-chat/Theme.css
