+++ examples/react/e-commerce-with-chat/assets/cover-mobile.png
