+++ examples/react/e-commerce-with-chat/assets/cover.png
