+++ examples/react/e-commerce-with-chat/assets/favicon.png
