+++ examples/react/e-commerce-with-chat/assets/manifest.webmanifest
