+++ examples/react/e-commerce-with-chat/components/AlgoliaSvg.tsx
