+++ examples/react/e-commerce-with-chat/components/ClearFilters.tsx
