+++ examples/react/e-commerce-with-chat/components/ClearFiltersMobile.tsx
