+++ examples/react/e-commerce-with-chat/components/NoResults.tsx
