+++ examples/react/e-commerce-with-chat/components/NoResultsBoundary.tsx
