+++ examples/react/e-commerce-with-chat/components/Panel.tsx
