+++ examples/react/e-commerce-with-chat/components/PriceSlider.css
