+++ examples/react/e-commerce-with-chat/components/Ratings.tsx
