+++ examples/react/e-commerce-with-chat/components/ResultsNumberMobile.tsx
