+++ examples/react/e-commerce-with-chat/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "example-react-instantsearch-e-commerce-vite",
+  "name": "example-react-instantsearch-e-commerce-with-chat",
   "version": "66.0.0",
   "private": true,
   "scripts": {
     "build": "vite build",
-    "website:examples": "vite build --base=. --outDir=../../../website/examples/react/e-commerce-vite",
+    "website:examples": "vite build --base=. --outDir=../../../website/examples/react/e-commerce-with-chat",
     "start": "vite",
     "dev": "vite",
     "preview": "vite preview"
@@ -28,6 +28,6 @@
     "@types/react-dom": "19.0.3",
     "@vitejs/plugin-react": "^4.0.0",
     "typescript": "5.5.2",
-    "vite": "^4.0.0"
+    "vite": "5.0.7"
   }
 }
