+++ examples/react/e-commerce-with-chat/utils/index.ts
