+++ examples/react/e-commerce-with-chat/vite.config.ts
