--- packages/react-instantsearch/src/widgets/Chat/index.ts
@@ -1,1 +1,0 @@
-export * from './Chat';
